backend/requirements.txt
# Core Flask
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-PyMongo==2.3.0

# Machine Learning
tensorflow==2.13.0
scikit-learn==1.3.0
joblib==1.3.1
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Image Processing
Pillow==10.0.0
opencv-python==4.8.0.74
opencv-contrib-python==4.8.0.74

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2

# API and HTTP
requests==2.31.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.7  # For PostgreSQL
pymongo==4.4.0  # For MongoDB

# Utilities
tqdm==4.65.0
pyyaml==6.0
schedule==1.2.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Production
gunicorn==21.2.0
gevent==23.7.0